rename: auth 패키지 생성 (#22)
