+++ 4-todo-jpa/todo/src/main/java/com/example/todo/auth/dto/LoginUserRequest.java
@@ -1,4 +1,4 @@
-package com.example.todo.user.dto;
+package com.example.todo.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
